refactor: database schema management locally
