+++ src/main/java/org/example/learningsystem/multitenancy/db/datasource/LocalTenantDataSourceProvider.java
@@ -1,6 +1,7 @@
 package org.example.learningsystem.multitenancy.db.datasource;
 
 import lombok.RequiredArgsConstructor;
+import org.example.learningsystem.multitenancy.db.schema.LocalSchemaHelper;
 import org.example.learningsystem.multitenancy.context.TenantInfo;
 import org.example.learningsystem.multitenancy.db.schema.TenantSchemaResolver;
 import org.example.learningsystem.multitenancy.db.util.TenantSchemaUtils;
@@ -8,12 +9,8 @@ import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
 import javax.sql.DataSource;
-import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
-import static org.example.learningsystem.multitenancy.constant.SqlConstants.SELECT_SCHEMAS_SQL;
-import static org.example.learningsystem.core.db.util.DatabaseUtils.queryForStringList;
 
 /**
  * Local implementation of {@link TenantDataSourceProvider} that uses a single shared data source for all tenants,
@@ -25,6 +22,7 @@ import static org.example.learningsystem.core.db.util.DatabaseUtils.queryForStri
 public class LocalTenantDataSourceProvider implements TenantDataSourceProvider {
 
     private final DataSource dataSource;
+    private final LocalSchemaHelper localSchemaHelper;
     private final TenantSchemaResolver tenantSchemaResolver;
 
     @Override
@@ -34,16 +32,10 @@ public class LocalTenantDataSourceProvider implements TenantDataSourceProvider {
 
     @Override
     public Map<TenantInfo, DataSource> getAll() {
-        var schemas = getSchemaNames(dataSource);
-        return schemas.stream()
-                .collect(Collectors.toMap(this::resolveTenantInfo, schema -> dataSource));
-    }
-
-    private List<String> getSchemaNames(DataSource dataSource) {
-        var names = queryForStringList(dataSource, SELECT_SCHEMAS_SQL);
+        var names = localSchemaHelper.getAll();
         return names.stream()
                 .filter(TenantSchemaUtils::isTenantSchema)
-                .toList();
+                .collect(Collectors.toMap(this::resolveTenantInfo, schema -> dataSource));
     }
 
     private TenantInfo resolveTenantInfo(String schema) {
